pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "road-segmentation"
version = "0.1.0"
description = "Road segmentation project scaffold for the Hudhud take-home assignment."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "kagglehub>=0.3.3",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "Pillow>=10.0.0",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "segmentation-models-pytorch>=0.3.3",
  "albumentations>=1.3.1",
  "opencv-python-headless>=4.8.0",
  "scikit-learn>=1.3.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "httpx>=0.25.0",
  "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
